Improved stability by adding mutex to secure threads working on the same things
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@98232 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 9784d7255c33122ff59ca35ca7929b92c9c84442

+++ Online/Online/confDBsynchronisationQt4v1/src/controller/mainwindowcontroller.py
@@ -19,16 +19,14 @@ from tfcmunin01controller import TFCMunin01Controller
 from showlogwindowcontroller import ShowLogWindowController
 from selectlogfilecontroller import SelectLogFileController
 from createsummarylogcontroller import CreateSummaryLogController
-from connectcontroller import ConnectController
 from deletedeviceswindowcontroller import DeleteDevicesWindowController
 
 class MainWindowController(Controller):
     def __init__(self, application):
         Controller.__init__(self, True)
         self.application = application
+        self.initialiseMutex()
         self.equipDBSystem = None
-        self.sparesMutex = QMutex()
-        self.devicesMutex = QMutex()
         self.devicesInfo = self.getDevicesInfo()
         self.sparesInfo = self.getSparesInfo()
         self.mainWindow = MainWindow(self)
@@ -36,6 +34,13 @@ class MainWindowController(Controller):
         self.sparesWidget = SparesWidget(self.sparesInfo, self.mainWindow, self)
         self.connectActions(self.mainWindow)
         self.mainWindow.show()
+    def initialiseMutex(self):
+        self.sparesMutex = QMutex()
+        self.devicesMutex = QMutex()
+        self.showLogMutex = QMutex()
+        self.refreshMutex = QMutex()
+        self.devicesActionMutex = QMutex()
+        self.sparesActionMutex = QMutex()
     def getDevicesInfo(self):
         confDB = self.getConfDBCxOracleConnection()
         equipDB = self.getEquipDBCxOracleConnection()
@@ -99,43 +104,54 @@ class MainWindowController(Controller):
         print "MainWindowController.onSelectLog() end"
     def onShowLastSessionLog(self):
         print "MainWindowController.onShowLastSessionLog() start"
-        self.logfile.flush()
-        self.showLogWindowController = ShowLogWindowController(self, False)
+        if self.showLogMutex.tryLock():
+            self.logfile.flush()
+            self.showLogWindowController = ShowLogWindowController(self, False)
+            self.showLogMutex.unlock()
         print "MainWindowController.onShowLastSessionLog() end"
     def onShowLastSummaryLog(self):
         print "MainWindowController.onShowLastSummaryLog() start"
-        self.logfile.flush()
-        self.showLogWindowController = ShowLogWindowController(self, True)
+        if self.showLogMutex.tryLock():
+            self.logfile.flush()
+            self.showLogWindowController = ShowLogWindowController(self, True)
+            self.showLogMutex.unlock()
         print "MainWindowController.onShowLastSummaryLog() end"
     def onInsert(self):
         print "MainWindowController.onInsert() start"
-        confDB = self.getConfDBCxOracleConnection()
-        equipDB = self.getEquipDBCxOracleConnection()
-        self.newDevicesWindowController = NewDevicesWindowController(confDB, equipDB, self)
+        if self.devicesActionMutex.tryLock():
+            confDB = self.getConfDBCxOracleConnection()
+            equipDB = self.getEquipDBCxOracleConnection()
+            self.newDevicesWindowController = NewDevicesWindowController(self)
         print "MainWindowController.onInsert() end"
     def onRefreshDevices(self):
         print "MainWindowController.onRefreshDevices() start"
-        self.logfile.flush()
-        self.refreshWorker = RefreshWorker(self.getEquipDBCxOracleConnection(), self.getConfDBCxOracleConnection(), self.spareDB, self.devicesInfo, None, self)
-        self.refreshWorker.start()
+        if self.refreshMutex.tryLock():
+            self.logfile.flush()
+            self.refreshWorker = RefreshWorker(self.getEquipDBCxOracleConnection(), self.getConfDBCxOracleConnection(), self.spareDB, self.devicesInfo, None, self)
+            self.refreshWorker.start()
+            self.refreshMutex.unlock()
         print "MainWindowController.onRefreshDevices() end"
     def onRefreshSpares(self):
         print "MainWindowController.onRefreshSpares() start"
-        self.logfile.flush()
-        self.refreshWorker = RefreshWorker(self.getEquipDBCxOracleConnection(), self.getConfDBCxOracleConnection(), self.spareDB, None, self.sparesInfo, self)
-        self.refreshWorker.start()
+        if self.refreshMutex.tryLock():
+            self.logfile.flush()
+            self.refreshWorker = RefreshWorker(self.getEquipDBCxOracleConnection(), self.getConfDBCxOracleConnection(), self.spareDB, None, self.sparesInfo, self)
+            self.refreshWorker.start()
+            self.refreshMutex.unlock()
         print "MainWindowController.onRefreshSpares() end"
     def onRefresh(self):
         print "MainWindowController.onRefresh() start"
-        self.logfile.flush()
-        self.refreshWorker = RefreshWorker(self.getEquipDBCxOracleConnection(), self.getConfDBCxOracleConnection(), self.spareDB, self.devicesInfo, self.sparesInfo, self)
-        self.refreshWorker.start()
+        if self.refreshMutex.tryLock():
+            self.logfile.flush()
+            self.refreshWorker = RefreshWorker(self.getEquipDBCxOracleConnection(), self.getConfDBCxOracleConnection(), self.spareDB, self.devicesInfo, self.sparesInfo, self)
+            self.refreshWorker.start()
+            self.refreshMutex.unlock()
         print "MainWindowController.onRefresh() end"
     def onRefreshDevicesWidget(self):
         print "MainWindowController.onRefreshDevicesWidget() start"
-        self.devicesMutex.lock()
-        self.devicesWidget.setDevicesInfo(self.devicesInfo)
-        self.devicesMutex.unlock()
+        if self.refreshMutex.tryLock():
+            self.devicesWidget.setDevicesInfo(self.devicesInfo)
+            self.refreshMutex.unlock()
         print "MainWindowController.onRefreshDevicesWidget() end"
     def onRefreshSparesWidget(self):
         print "MainWindowController.onRefreshSparesWidget() start"
@@ -145,36 +161,44 @@ class MainWindowController(Controller):
         print "MainWindowController.onRefreshSparesWidget() end"
     def onUpdate(self):
         print "MainWindowController.onUpdate() start"
-        self.updateDevicesController = UpdateDevicesController(self)
+        if self.devicesActionMutex.tryLock():
+            self.updateDevicesController = UpdateDevicesController(self)
         print "MainWindowController.onUpdate() end"
     def onDelete(self):
         print "MainWindowController.onDelete() start"
-        self.db = self.getCn()
-        self.deleteDevicesWindowController = DeleteDevicesWindowController(self)
+        if self.devicesActionMutex.tryLock():
+            self.db = self.getCn()
+            self.deleteDevicesWindowController = DeleteDevicesWindowController(self)
         print "MainWindowController.onDelete() end"
     def onConnect(self):
         print "MainWindowController.onConnect() start"
-        self.db = self.getCn()
-        self.connectController = ConnectController(self)
+        if self.connectMutex.tryLock():
+            self.db = self.getCn()
+            self.connectMutex.unlock()
         print "MainWindowController.onConnect() end"
     def onTfcmunin01(self):
         print "MainWindowController.onTfcmunin01() start"
-        self.tFCMunin01Controller = TFCMunin01Controller(self)
+        if self.devicesActionMutex.tryLock():
+            self.tFCMunin01Controller = TFCMunin01Controller(self)
         print "MainWindowController.onTfcmunin01() end"
     def onSwap(self):
         print "MainWindowController.onSwap() start"
-        self.swapDevicesWindowController = SwapDevicesWindowController(self)
+        if self.devicesActionMutex.tryLock() and self.sparesActionMutex.tryLock():
+            self.swapDevicesWindowController = SwapDevicesWindowController(self)
         print "MainWindowController.onSwap() end"
     def onInsertNewSpares(self):
         print "MainWindowController.onInsertNewSpares() start"
-        self.insertNewSparesController = InsertNewSparesController(self)
+        if self.sparesActionMutex.tryLock():
+            self.insertNewSparesController = InsertNewSparesController(self)
         print "MainWindowController.onInsertNewSpares() end"
     def onUpdateSpares(self):
         print "MainWindowController.onUpdateSpares() start"
-        self.updateNewSparesController = UpdateNewSparesController(self)
+        if self.sparesActionMutex.tryLock():
+            self.updateNewSparesController = UpdateNewSparesController(self)
         print "MainWindowController.onUpdateSpares() end"
     def onDeleteSpares(self):
         print "MainWindowController.onDeleteSpares() start"
-        self.deleteSparesController = DeleteSparesController(self)
+        if self.sparesActionMutex.tryLock():
+            self.deleteSparesController = DeleteSparesController(self)
         print "MainWindowController.onDeleteSpares() end"
